Problem: need to know if zyre_start() failed
Following #714d80 in CZMQ, it's now possible to detect when the
current system does not support beacons. Now we need to know at
the Zyre app level if beaconing is broken.

Solution: expand zyre_start() to return a status code, 0 if OK,
-1 if beaconing is broken. We use this in the selftests to skip
if the system looks broken.

Test case is a laptop with WiFi switched off, and no ethernet, as
the loopback interface does not do UDP broadcasts and thus cannot
work for beaconing.