Problem: nodes don't rejoin properly after reconnect
There is something weird going on in libzmq, I think, and this still
needs to be investigated. Node A DEALER connects to node B ROUTER and
sends messages once a second, and then the network interface dies, and
then node A destroys its socket. Then it creates a new socket, connects
to same endpoint B, and sends one new message. The old messages do
arrive, but the new one does not.

Solution: (and this is a workaround), remove the peer if we get an
unexpected command from it. See zyre_node.c:599.

Fixes #200.