Set ZMQ_ROUTER_HANDOVER on node inbox
Problem: was incorrectly taking old PING messages from disconnected
peer socket, and ignoring new HELLO commands. The previous fix was
to destroy the peer and retry. However that didn't address the real
problem.

Solution: set ZMQ_ROUTER_HANDOVER on the router socket, so that new
client connections always replace existing ones.

This appears to work properly; the test is to run several instances
of zpinger on a laptop connected to a WiFi network, then disable the
WiFi interface, and then re-enable it. zpinger should first report
that its peers have exited, and then that they reappear.