Zyre logging system removed
This is a substantial change. The logging system has been removed,
firstly because it did not work as well as hoped, and second it
made Zyre more complex than necessary. If we need a ZeroMQ-based
logging system we should add it as a separate protocol. Zyre now
uses the zsys_xxx logging API, to send all log output to stdout
or to syslog (if enabled by the application).