Problem: randomly claiming "lost messages" after reconnects
The sequence detection logic was not correct, and when receiving a
HELLO from a disconnected/reconnected peer, would complain that it
did not expect the sequence number (1).

Solution: clarify in RFC 36 that HELLO always has sequence = 1.

Then, when receiving a message, if it's HELLO, reset expected
sequence to 1.

Lastly, when getting a bad sequence, delete the peer.