Cleaned-up endpoint announcement
In gossip model, caller specifies bind/connect endpoints explicitly
both for the ZRE protocol and for the zgossip protocol. The API now
tries to explain these better, and provide a method to set the
announced endpoint for ZRE, when it's not the same as the bind
value. Note that zbeacon does all this behind the scenes...