Problem: Difficult to determine when to stop processing events
Solution: Add a STOP event send from the local node in response to
zyre_stop.

Motivation: A common pattern when using Zyre is to have a background
thread process zyre events, and for the main thread to control program
execution. If we want to terminate the program, we must create a
separate channel to communicate the stop signal to the background
thread, which means we cannot use a simple while loop around zyre_recv
or zyre_event_new, complicating processing unnecessarily. Now we simply
listen for the stop event over the same channel, and terminate event
processing gracefully.