Problem: hellozyre is looking for bindings module
It doesn't need this. The bindings module is useful only when loading
a binding from the directory where you're developing it. Thus it looks
in places like build/Release. In applications that use a binding, we
don't have this directory structure.

Solution: expect zyre.node in node_modules (default location)